Move interface() to libImportOncokb.py
PR feedback Hongxin
